move password_strength and related function in common.py
