+++ src/add_users.py
@@ -1,16 +1,11 @@
 #!/usr/bin/env python
 
 from gi.repository import Gtk, Gdk
-import os
-import re
 import pickle
+from gbi_common import password_strength
 
 # Directory use from the installer.
 tmp = "/tmp/.gbi/"
-installer = "/usr/local/lib/gbi/"
-query = "sh /usr/local/lib/gbi/backend-query/"
-if not os.path.exists(tmp):
-    os.makedirs(tmp)
 userfile = tmp + "user"
 
 cssProvider = Gtk.CssProvider()
@@ -24,44 +19,8 @@ styleContext.add_provider_for_screen(
 )
 
 
-# Find if pasword contain only lower case and number
-def lowerCase(strg, search=re.compile(r'[^a-z]').search):
-    return not bool(search(strg))
-
-
-# Find if pasword contain only upper case
-def upperCase(strg, search=re.compile(r'[^A-Z]').search):
-    return not bool(search(strg))
-
-
-# Find if pasword contain only lower case and number
-def lowerandNunber(strg, search=re.compile(r'[^a-z0-9]').search):
-    return not bool(search(strg))
-
-
-# Find if pasword contain only upper case and number
-def upperandNunber(strg, search=re.compile(r'[^A-Z0-9]').search):
-    return not bool(search(strg))
-
-
-# Find if pasword contain only lower and upper case and
-def lowerUpperCase(strg, search=re.compile(r'[^a-zA-Z]').search):
-    return not bool(search(strg))
-
-
-# Find if pasword contain only lower and upper case and
-def lowerUpperNumber(strg, search=re.compile(r'[^a-zA-Z0-9]').search):
-    return not bool(search(strg))
-
-
-# Find if password contain only lowercase, uppercase numbers
-# and some special character.
-def allCharacter(strg):
-    search = re.compile(r'[^a-zA-Z0-9~\!@#\$%\^&\*_\+":;\'\-]').search
-    return not bool(search(strg))
-
-
 class AddUsers:
+
     def save_selection(self):
         f = open(userfile, 'wb')
         uname = self.user.get_text()
@@ -177,74 +136,9 @@ class AddUsers:
     def get_model(self):
         return self.vbox1
 
-    def password_strength(self):
-        password = self.password.get_text()
-        same_character_type = any(
-            [
-                lowerCase(password),
-                upperCase(password),
-                password.isdigit()
-            ]
-        )
-        mix_character = any(
-            [
-                lowerandNunber(password),
-                upperandNunber(password),
-                lowerUpperCase(password)
-            ]
-        )
-        if ' ' in password or '\t' in password:
-            self.label3.set_text("Space not allowed")
-        elif len(password) <= 4:
-            self.label3.set_text("Super Weak")
-        elif len(password) <= 8 and same_character_type:
-            self.label3.set_text("Super Weak")
-        elif len(password) <= 8 and mix_character:
-            self.label3.set_text("Very Weak")
-        elif len(password) <= 8 and lowerUpperNumber(password):
-            self.label3.set_text("Fairly Weak")
-        elif len(password) <= 8 and allCharacter(password):
-            self.label3.set_text("Weak")
-        elif len(password) <= 12 and same_character_type:
-            self.label3.set_text("Very Weak")
-        elif len(password) <= 12 and mix_character:
-            self.label3.set_text("Fairly Weak")
-        elif len(password) <= 12 and lowerUpperNumber(password):
-            self.label3.set_text("Weak")
-        elif len(password) <= 12 and allCharacter(password):
-            self.label3.set_text("Strong")
-        elif len(password) <= 16 and same_character_type:
-            self.label3.set_text("Fairly Weak")
-        elif len(password) <= 16 and mix_character:
-            self.label3.set_text("Weak")
-        elif len(password) <= 16 and lowerUpperNumber(password):
-            self.label3.set_text("Strong")
-        elif len(password) <= 16 and allCharacter(password):
-            self.label3.set_text("Fairly Strong")
-        elif len(password) <= 20 and same_character_type:
-            self.label3.set_text("Weak")
-        elif len(password) <= 20 and mix_character:
-            self.label3.set_text("Strong")
-        elif len(password) <= 20 and lowerUpperNumber(password):
-            self.label3.set_text("Fairly Strong")
-        elif len(password) <= 20 and allCharacter(password):
-            self.label3.set_text("Very Strong")
-        elif len(password) <= 24 and same_character_type:
-            self.label3.set_text("Strong")
-        elif len(password) <= 24 and mix_character:
-            self.label3.set_text("Fairly Strong")
-        elif len(password) <= 24 and lowerUpperNumber(password):
-            self.label3.set_text("Very Strong")
-        elif len(password) <= 24 and allCharacter(password):
-            self.label3.set_text("Super Strong")
-        elif same_character_type:
-            self.label3.set_text("Fairly Strong")
-        else:
-            self.label3.set_text("Super Strong")
-
     def password_verification(self, widget, button3):
-        self.password_strength()
         password = self.password.get_text()
+        password_strength(password, self.label3)
         repassword = self.repassword.get_text()
         if password == repassword and password != "" and " " not in password:
             self.img.set_from_stock(Gtk.STOCK_YES, 5)
